refactor: migrate octree sample fixtures to geo_algorithms
